Improving configuration module
Substrate configuration moved to substrate module
Move Bip44/49/84 coin indexes, coins and configuration mapping to conf module
