+++ bip_utils/conf/bip49_coin_conf.py
@@ -38,6 +38,7 @@ BIP49_BTC_P2SH_NET_VER_TEST: bytes = b"\xc4"
 # Configuration for Bitcoin main net
 Bip49BitcoinMainNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Bitcoin", "BTC"),
+    coin_idx=0,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -48,6 +49,7 @@ Bip49BitcoinMainNet: BipCoinConf = BipCoinConf(
 # Configuration for Bitcoin test net
 Bip49BitcoinTestNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Bitcoin TestNet", "BTC"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_TEST,
@@ -59,6 +61,7 @@ Bip49BitcoinTestNet: BipCoinConf = BipCoinConf(
 # Configuration for Bitcoin Cash main net
 Bip49BitcoinCashMainNet: BipBitcoinCashConf = BipBitcoinCashConf(
     coin_name=CoinNames("Bitcoin Cash", "BCH"),
+    coin_idx=145,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -70,6 +73,7 @@ Bip49BitcoinCashMainNet: BipBitcoinCashConf = BipBitcoinCashConf(
 # Configuration for Bitcoin Cash test net
 Bip49BitcoinCashTestNet: BipBitcoinCashConf = BipBitcoinCashConf(
     coin_name=CoinNames("Bitcoin Cash TestNet", "BCH"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_TEST,
@@ -82,6 +86,7 @@ Bip49BitcoinCashTestNet: BipBitcoinCashConf = BipBitcoinCashConf(
 # Configuration for BitcoinSV main net
 Bip49BitcoinSvMainNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("BitcoinSV", "BSV"),
+    coin_idx=236,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -92,6 +97,7 @@ Bip49BitcoinSvMainNet: BipCoinConf = BipCoinConf(
 # Configuration for BitcoinSV test net
 Bip49BitcoinSvTestNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("BitcoinSV TestNet", "BSV"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_TEST,
@@ -103,6 +109,7 @@ Bip49BitcoinSvTestNet: BipCoinConf = BipCoinConf(
 # Configuration for Dash main net
 Bip49DashMainNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Dash", "DASH"),
+    coin_idx=5,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -113,6 +120,7 @@ Bip49DashMainNet: BipCoinConf = BipCoinConf(
 # Configuration for Dash test net
 Bip49DashTestNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Dash TestNet", "DASH"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_TEST,
@@ -124,6 +132,7 @@ Bip49DashTestNet: BipCoinConf = BipCoinConf(
 # Configuration for Dogecoin main net
 Bip49DogecoinMainNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Dogecoin", "DOGE"),
+    coin_idx=3,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=KeyNetVersions(b"02facafd", b"02fac398"),   # dgub / dgpv
@@ -134,6 +143,7 @@ Bip49DogecoinMainNet: BipCoinConf = BipCoinConf(
 # Configuration for Dogecoin test net
 Bip49DogecoinTestNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Dogecoin TestNet", "DOGE"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=KeyNetVersions(b"0432a9a8", b"0432a243"),   # tgub / tgpv
@@ -145,6 +155,7 @@ Bip49DogecoinTestNet: BipCoinConf = BipCoinConf(
 # Configuration for Litecoin main net
 Bip49LitecoinMainNet: BipLitecoinConf = BipLitecoinConf(
     coin_name=CoinNames("Litecoin", "LTC"),
+    coin_idx=2,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -156,6 +167,7 @@ Bip49LitecoinMainNet: BipLitecoinConf = BipLitecoinConf(
 # Configuration for Litecoin test net
 Bip49LitecoinTestNet: BipLitecoinConf = BipLitecoinConf(
     coin_name=CoinNames("Litecoin TestNet", "LTC"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=KeyNetVersions(b"0436f6e1", b"0436ef7d"),       # ttub / ttpv
@@ -168,6 +180,7 @@ Bip49LitecoinTestNet: BipLitecoinConf = BipLitecoinConf(
 # Configuration for Zcash main net
 Bip49ZcashMainNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Zcash", "ZEC"),
+    coin_idx=133,
     is_testnet=False,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_MAIN,
@@ -178,6 +191,7 @@ Bip49ZcashMainNet: BipCoinConf = BipCoinConf(
 # Configuration for Zcash test net
 Bip49ZcashTestNet: BipCoinConf = BipCoinConf(
     coin_name=CoinNames("Zcash TestNet", "ZEC"),
+    coin_idx=1,
     is_testnet=True,
     def_path=NOT_HARDENED_DEF_PATH,
     key_net_ver=BIP49_BTC_KEY_NET_VER_TEST,
