+++ bip_utils/substrate/conf/__init__.py
@@ -1,0 +1,3 @@
+from bip_utils.substrate.conf.substrate_coins import SubstrateCoins
+from bip_utils.substrate.conf.substrate_coin_conf import SubstrateCoinConf
+from bip_utils.substrate.conf.substrate_conf_getter import SubstrateConfGetter
